Change MainActivity to fragment
Change MainActivity to fragment
