+++ koin/src/main/res/navigation/main_navigation_graph.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_navigation_graph"
+    app:startDestination="@id/main_fragment">
+
+
+    <fragment
+        android:id="@+id/main_fragment"
+        android:name="in.koreatech.koin.ui.main.MainFragment"
+        android:label="Blank"
+        tools:layout="@layout/fragment_main">
+
+        <action
+            android:id="@+id/navi_item_circle_action"
+            app:destination="@+id/circle_main_fragment" />
+    </fragment>
+
+
+    <fragment
+        android:id="@+id/circle_main_fragment"
+        android:name="in.koreatech.koin.ui.main.MainFragment"
+        android:label="Blank"
+        tools:layout="@layout/fragment_main">
+
+    </fragment>
+
+</navigation>
